Bug: Dans la fonction file_empty, il faut que le fichier existe pour pouvoir le vider.
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@12008 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPfiles/classes/files.class.php
@@ -379,7 +379,9 @@ class CFile extends CDocumentItem {
   }
 	
   function file_empty() {
-    file_put_contents($this->_file_path, '');
+    if (file_exists($this->_file_path)) {
+      file_put_contents($this->_file_path, '');
+    }
   }
   
   function isPDFconvertible() {
